fix: refactor worker runtime stats to model boundary
