+++ src/main/kotlin/io/github/cdsap/testprocess/agent/WorkerRegistry.kt
@@ -1,6 +1,7 @@
 package io.github.cdsap.testprocess.agent
 
 import io.github.cdsap.testprocess.model.TestProcess
+import io.github.cdsap.testprocess.model.WorkerRuntimeStats
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import java.io.File
@@ -27,23 +28,6 @@ data class WorkerRegistryEntry(
         else -> "${bytes}b"
     }
 }
-
-@Serializable
-data class WorkerRuntimeStats(
-    val pid: Long,
-    val uptimeMs: Long = 0,
-    val cpuTimeMs: Long = -1,
-    val usedHeapBytes: Long = 0,
-    val peakHeapBytes: Long = 0,
-    val peakMetaspaceBytes: Long = 0,
-    val maxHeapBytes: Long = 0,
-    val gcCollections: Long = 0,
-    val gcTimeMs: Long = 0,
-    val gcType: String = "Unknown",
-    val jitTimeMs: Long = -1,
-    val classesLoaded: Long = -1,
-    val peakThreads: Int = -1
-)
 
 object WorkerRegistry {
     private val json = Json { ignoreUnknownKeys = true }
