+++ src/main/kotlin/io/github/cdsap/testprocess/model/PersistedState.kt
@@ -1,6 +1,5 @@
 package io.github.cdsap.testprocess.model
 
-import io.github.cdsap.testprocess.agent.WorkerRuntimeStats
 import kotlinx.serialization.Serializable
 
 @Serializable
